# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Quant Trading System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StrategyType = typeof StrategyType[keyof typeof StrategyType];


export const StrategyType = {
  ma_crossover: 'ma_crossover',
  rsi: 'rsi',
  bollinger: 'bollinger',
  grid: 'grid',
} as const;

export type StrategyParams = { [key: string]: unknown };

export interface Strategy {
  id: number;
  name: string;
  type: StrategyType;
  symbol: string;
  interval: string;
  params: StrategyParams;
  isActive: boolean;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export type StrategyInputType = typeof StrategyInputType[keyof typeof StrategyInputType];


export const StrategyInputType = {
  ma_crossover: 'ma_crossover',
  rsi: 'rsi',
  bollinger: 'bollinger',
  grid: 'grid',
} as const;

export type StrategyInputParams = { [key: string]: unknown };

export interface StrategyInput {
  name: string;
  type: StrategyInputType;
  symbol: string;
  interval: string;
  params: StrategyInputParams;
  description?: string;
}

export type StrategyUpdateParams = { [key: string]: unknown };

export interface StrategyUpdate {
  name?: string;
  symbol?: string;
  interval?: string;
  params?: StrategyUpdateParams;
  isActive?: boolean;
  description?: string;
}

export type BacktestResultStatus = typeof BacktestResultStatus[keyof typeof BacktestResultStatus];


export const BacktestResultStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface EquityPoint {
  date: string;
  value: number;
}

export type BacktestTradeSide = typeof BacktestTradeSide[keyof typeof BacktestTradeSide];


export const BacktestTradeSide = {
  long: 'long',
  short: 'short',
} as const;

export interface BacktestTrade {
  id: number;
  side: BacktestTradeSide;
  entryPrice: number;
  exitPrice: number;
  entryDate: string;
  exitDate: string;
  pnl: number;
  pnlPct: number;
}

export interface BacktestResult {
  id: number;
  strategyId: number;
  strategyName: string;
  symbol: string;
  interval: string;
  startDate: string;
  endDate: string;
  initialCapital: number;
  finalCapital: number;
  totalReturn: number;
  annualizedReturn: number;
  maxDrawdown: number;
  winRate: number;
  profitFactor: number;
  totalTrades: number;
  status: BacktestResultStatus;
  /** @nullable */
  equityCurve?: EquityPoint[] | null;
  /** @nullable */
  trades?: BacktestTrade[] | null;
  createdAt: string;
}

export interface BacktestInput {
  strategyId: number;
  startDate: string;
  endDate: string;
  initialCapital: number;
}

export type PositionSide = typeof PositionSide[keyof typeof PositionSide];


export const PositionSide = {
  long: 'long',
  short: 'short',
} as const;

export interface Position {
  id: number;
  symbol: string;
  side: PositionSide;
  size: number;
  entryPrice: number;
  currentPrice: number;
  unrealizedPnl: number;
  unrealizedPnlPct: number;
  strategyName: string;
  openedAt: string;
}

export type OrderSide = typeof OrderSide[keyof typeof OrderSide];


export const OrderSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type OrderType = typeof OrderType[keyof typeof OrderType];


export const OrderType = {
  market: 'market',
  limit: 'limit',
  stop: 'stop',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  filled: 'filled',
  cancelled: 'cancelled',
  rejected: 'rejected',
} as const;

export interface Order {
  id: number;
  symbol: string;
  side: OrderSide;
  type: OrderType;
  size: number;
  price: number;
  status: OrderStatus;
  strategyName: string;
  createdAt: string;
}

export interface PaperTradingSummary {
  totalEquity: number;
  availableBalance: number;
  totalUnrealizedPnl: number;
  totalRealizedPnl: number;
  totalTrades: number;
  winRate: number;
  activeStrategies: number;
}

export interface StrategyExecution {
  strategyId: number;
  isRunning: boolean;
}

export interface RiskSettings {
  id: number;
  maxPositionSizePct: number;
  maxDailyLossPct: number;
  maxLeverage: number;
  maxOpenPositions: number;
  stopLossPct: number;
  takeProfitPct: number;
  trailingStopPct: number;
  trailingStopEnabled: boolean;
  emergencyStopEnabled: boolean;
  updatedAt: string;
}

export interface RiskSettingsUpdate {
  maxPositionSizePct?: number;
  maxDailyLossPct?: number;
  maxLeverage?: number;
  maxOpenPositions?: number;
  stopLossPct?: number;
  takeProfitPct?: number;
  trailingStopPct?: number;
  trailingStopEnabled?: boolean;
  emergencyStopEnabled?: boolean;
}

export interface EmergencyStopResult {
  success: boolean;
  stoppedStrategies: number;
  message: string;
}

export type TradeRecordSide = typeof TradeRecordSide[keyof typeof TradeRecordSide];


export const TradeRecordSide = {
  buy: 'buy',
  sell: 'sell',
} as const;

export type TradeRecordMode = typeof TradeRecordMode[keyof typeof TradeRecordMode];


export const TradeRecordMode = {
  paper: 'paper',
  live: 'live',
} as const;

export interface TradeRecord {
  id: number;
  symbol: string;
  side: TradeRecordSide;
  size: number;
  price: number;
  fee: number;
  pnl: number;
  pnlPct: number;
  strategyName: string;
  executedAt: string;
  mode: TradeRecordMode;
}

export interface MarketTicker {
  symbol: string;
  price: number;
  change24h: number;
  changePct24h: number;
  volume24h: number;
  high24h: number;
  low24h: number;
}

export type DashboardStatsSystemMode = typeof DashboardStatsSystemMode[keyof typeof DashboardStatsSystemMode];


export const DashboardStatsSystemMode = {
  paper: 'paper',
  live: 'live',
} as const;

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
}

export interface DashboardStats {
  totalStrategies: number;
  activeStrategies: number;
  totalBacktests: number;
  totalTrades: number;
  paperTradingPnl: number;
  paperTradingPnlPct: number;
  systemMode: DashboardStatsSystemMode;
  recentActivity: ActivityItem[];
}

export type SystemSettingsMode = typeof SystemSettingsMode[keyof typeof SystemSettingsMode];


export const SystemSettingsMode = {
  paper: 'paper',
  live: 'live',
} as const;

export interface SystemSettings {
  id: number;
  mode: SystemSettingsMode;
  exchangeName: string;
  apiKeySet: boolean;
  tradingEnabled: boolean;
  updatedAt: string;
}

export type SystemSettingsUpdateMode = typeof SystemSettingsUpdateMode[keyof typeof SystemSettingsUpdateMode];


export const SystemSettingsUpdateMode = {
  paper: 'paper',
  live: 'live',
} as const;

export interface SystemSettingsUpdate {
  mode?: SystemSettingsUpdateMode;
  exchangeName?: string;
}