# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Crypto Quant Trading System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List all strategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a new strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get a strategy by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "404":
          description: Not found
    patch:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /backtests:
    get:
      operationId: listBacktests
      tags: [backtests]
      summary: List all backtest results
      responses:
        "200":
          description: List of backtest results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BacktestResult"
    post:
      operationId: runBacktest
      tags: [backtests]
      summary: Run a backtest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestInput"
      responses:
        "201":
          description: Backtest result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestResult"
  /backtests/{id}:
    get:
      operationId: getBacktest
      tags: [backtests]
      summary: Get backtest result by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Backtest detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestResult"
...
components:
  schemas:
...
    Strategy:
      type: object
      required: [id, name, type, symbol, interval, params, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [ma_crossover, rsi, bollinger, grid]
        symbol:
          type: string
        interval:
          type: string
        params:
          type: object
          additionalProperties: true
        isActive:
          type: boolean
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    StrategyInput:
      type: object
      required: [name, type, symbol, interval, params]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [ma_crossover, rsi, bollinger, grid]
        symbol:
          type: string
        interval:
          type: string
        params:
          type: object
          additionalProperties: true
        description:
          type: string
    StrategyUpdate:
      type: object
      properties:
        name:
          type: string
        symbol:
          type: string
        interval:
          type: string
        params:
          type: object
          additionalProperties: true
        isActive:
          type: boolean
        description:
          type: string
    BacktestResult:
      type: object
      required: [id, strategyId, strategyName, symbol, interval, startDate, endDate, initialCapital, finalCapital, totalReturn, annualizedReturn, maxDrawdown, winRate, profitFactor, totalTrades, status, createdAt]
      properties:
        id:
          type: integer
        strategyId:
          type: integer
        strategyName:
          type: string
        symbol:
          type: string
        interval:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        initialCapital:
          type: number
        finalCapital:
          type: number
        totalReturn:
          type: number
        annualizedReturn:
          type: number
        maxDrawdown:
          type: number
        winRate:
          type: number
        profitFactor:
          type: number
        totalTrades:
          type: integer
        status:
          type: string
          enum: [pending, running, completed, failed]
        equityCurve:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/EquityPoint"
        trades:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/BacktestTrade"
        createdAt:
          type: string
    BacktestInput:
      type: object
      required: [strategyId, startDate, endDate, initialCapital]
      properties:
        strategyId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        initialCapital:
          type: number
    EquityPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number
    BacktestTrade:
      type: object
      required: [id, side, entryPrice, exitPrice, entryDate, exitDate, pnl, pnlPct]
      properties:
        id:
          type: integer
        side:
          type: string
          enum: [long, short]
        entryPrice:
          type: number
        exitPrice:
          type: number
        entryDate:
          type: string
        exitDate:
          type: string
        pnl:
          type: number
        pnlPct:
          type: number
...