# lib/api-zod/src/generated/types/backtestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Quant Trading System API
 * OpenAPI spec version: 0.1.0
 */
import type { BacktestResultStatus } from './backtestResultStatus';
import type { BacktestTrade } from './backtestTrade';
import type { EquityPoint } from './equityPoint';

export interface BacktestResult {
  id: number;
  strategyId: number;
  strategyName: string;
  symbol: string;
  interval: string;
  startDate: string;
  endDate: string;
  initialCapital: number;
  finalCapital: number;
  totalReturn: number;
  annualizedReturn: number;
  maxDrawdown: number;
  winRate: number;
  profitFactor: number;
  totalTrades: number;
  status: BacktestResultStatus;
  /** @nullable */
  equityCurve?: EquityPoint[] | null;
  /** @nullable */
  trades?: BacktestTrade[] | null;
  createdAt: string;
}