# lib/api-zod/src/generated/types/strategyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Quant Trading System API
 * OpenAPI spec version: 0.1.0
 */
import type { StrategyInputParams } from './strategyInputParams';
import type { StrategyInputType } from './strategyInputType';

export interface StrategyInput {
  name: string;
  type: StrategyInputType;
  symbol: string;
  interval: string;
  params: StrategyInputParams;
  description?: string;
}